+++ blockchain/chain/src/store/mod.rs
@@ -1,10 +1,12 @@
 // Copyright 2020 ChainSafe Systems
 // SPDX-License-Identifier: Apache-2.0, MIT
 
+mod base_fee;
 mod chain_store;
 mod errors;
 mod tip_index;
 
+pub use self::base_fee::*;
 pub use self::chain_store::*;
 pub use self::errors::*;
 pub use self::tip_index::*;
